requirements.txt
streamlit
gpt4all
pymupdf
pytesseract
pdf2image
pillow
